- IE.findInternetExplorer(constraint) now uses ShellWindows2 (instead of SHDocVw.ShelWindows). Fixes AttachToIE and IE.Exist and starting IE in new process on Vista and Win2008.
- Fixed issue with ElementFinder.RetrieveIdConstraint resulting in failing tests
- some clean ups of using

git-svn-id: svn://svn.code.sf.net/p/watin/code/trunk@812 6ca19b5b-dac0-4595-9ef0-4637b6ca95b2

+++ src/UnitTests/IETests/ShellWindows2Tests.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Diagnostics;
 using System.Threading;
 using NUnit.Framework;
 using NUnit.Framework.SyntaxHelpers;
+using SHDocVw;
 using WatiN.Core.Native.InternetExplorer;
 
 namespace WatiN.Core.UnitTests.IETests
@@ -10,38 +12,60 @@ namespace WatiN.Core.UnitTests.IETests
     public class ShellWindows2Tests
     {
         [Test]
-        public void ShouldHaveACountProperty()
+        public void ShouldCount1IEInstance()
         {
-            // GIVEN
-            var documents2 = new ShellWindows2();
+            Process process = null;
+            try
+            {
+                // GIVEN
+                process = StartIE("about:blank");
+                Assert.That(process, Is.Not.Null, "pre-condition: Expected an IE process");
+                
+                var browsers = new ShellWindows2();
 
-            // WHEN
-            var count = documents2.Count;
+                // WHEN
+                var count = browsers.Count;
 
-            // THEN
-            Assert.That(count, Is.EqualTo(0), "unexpected count");
+                foreach (IWebBrowser2 browser in browsers)
+                {
+                    Console.WriteLine(browser.LocationURL);
+                }
+                // THEN
+                Assert.That(count, Is.EqualTo(1), "unexpected count");
+            }
+            finally
+            {
+                if (process != null) process.Kill();
+            }
         }
 
         [Test]
-        public void ShouldCount1IEInstance()
+        public void ShouldBeAbleToEnumerateFoundInstances()
         {
-            var process = StartIE("about:blank");
-            Assert.That(process, Is.Not.Null, "pre-condition: Expected an IE process");
+            Process process1 = null;
+            Process process2 = null;
 
             try
             {
                 // GIVEN
+                process1 = StartIE("about:blank");
+                Assert.That(process1, Is.Not.Null, "pre-condition 1: Expected an IE process");
+                
+                process2 = StartIE("about:blank");
+                Assert.That(process2, Is.Not.Null, "pre-condition 2: Expected an IE process");
+
                 var documents2 = new ShellWindows2();
 
                 // WHEN
                 var count = documents2.Count;
 
                 // THEN
-                Assert.That(count, Is.EqualTo(1), "unexpected count");
+                Assert.That(count, Is.EqualTo(2), "unexpected count");
             }
             finally
             {
-                process.Kill();
+                if (process1 != null) process1.Kill();
+                if (process2 != null) process2.Kill();
             }
         }
 
@@ -50,22 +74,13 @@ namespace WatiN.Core.UnitTests.IETests
             var m_Proc = Process.Start("IExplore.exe", url);
             if (m_Proc == null) return null;
 
-            var retry = 10;
+            // This sleep is necesary to give IE time to fully instantiate.
+            // Needed sleep time might differ from machine to machine..
+            Thread.Sleep(2000);
 
-            do
-            {
-                retry--;
+            m_Proc.Refresh();
 
-                m_Proc.Refresh();
-                var mainWindowHandle = (int)m_Proc.MainWindowHandle;
-
-                if (mainWindowHandle != 0) { break; }
-
-                Thread.Sleep(500);
-
-            } while (retry >= 0);
-
-            return retry == 0 ? null : m_Proc;
+            return m_Proc;
         }
     }
 }
